tg_bot/requirements.txt
pyTelegramBotAPI
paho.mqtt